Deleted database.js and splitted functions into new files in ./database folder. Created jsonModels folder with example documents for db collections and response messages. Created express router files to handle rest api requests.
