+++ server.js
@@ -2,10 +2,8 @@ const express = require('express');
 const app = express();
 const bodyParser = require('body-parser');
 const MongoClient = require('mongodb').MongoClient;
-const database = require('./database');
+const userWidgetsCollectionUtils = require('./database/userWidgetsCollectionUtils');
 const jwt = require('jsonwebtoken');
-const multer = require('multer');
-const uuid = require('uuid/v4');
 const ObjectId = require('mongodb').ObjectId;
 
 const port = process.env.PORT || 5000;
@@ -15,257 +13,21 @@ var http = require('http').Server(app);
 var io = require('socket.io')(http);
 var shell = require('shelljs');
 
-const currentUser = "Emre";
+
+var apiRouter = require('./routes/api');
+var nativeRouter = require('./routes/native');
+
+const currentUser = "felix";
 const mongoURL = 'mongodb://127.0.0.1:27017/smartmirror';
 
 require('dotenv').load();
 
-//Setup Multer for uploading images
-const storage = multer.diskStorage({
-    destination: 'public/uploads/',
-    filename: function (req, file, callback) {
-        switch (file.mimetype) {
-            case 'image/jpeg':
-                ext = '.jpeg';
-                break;
-            case 'image/png':
-                ext = '.png';
-                break;
-            default:
-                ext = '';
-        }
-        callback(null, uuid() + ext);
-    }
-});
-const upload = multer({storage: storage});
-
-
-// for jsonwebtoken and session, verifies session token
-function verifyToken(req, res, next) {
-    const bearerHeader = req.headers.authorization;
-    if (typeof bearerHeader !== 'undefined') {
-        const bearer = bearerHeader.split(' ');
-        const bearerToken = bearer[1];
-        req.token = bearerToken;
-        next();
-    } else {
-        res.json({
-            message: "User is not authorized"
-        });
-    }
-}
 
 // Body parser to decode incoming json
 app.use(bodyParser.json());
 
-app.get('/api/hello', (req, res) => {
-    res.send({express: 'Hello From Express'});
-});
-
-app.get('/api/user/getUserWidgets', (req, res) => {
-    MongoClient.connect(mongoURL, {useNewUrlParser: true}, function (err, client) {
-        if (err) {
-            console.log('Unable to connect to MongoDB');
-            throw err;
-        } else {
-            database.processGetUserWidgets(client.db('smartmirror'), req.query.user_id, res, client);
-        }
-    });
-});
-
-// HTTP Requests
-// check if token is authorized
-app.post('/native/authizeToken', verifyToken, (req, res) => {
-    jwt.verify(req.token, process.env.secretkey, (err, authData) => {
-        console.log("Checking auth token");
-        if (err) {
-            console.log("User not verified");
-            res.send(JSON.stringify({
-                authorized: false,
-                message: "Token not authorized. Please login!"
-            }));
-        } else {
-            console.log("User is verified");
-            res.send(JSON.stringify({
-                authorized: true,
-                message: "Token is authorized. All good!"
-            }));
-        }
-    });
-});
-
-// Register, check user credentials and create user with jwt
-app.post('/native/signup', (req, res) => {
-    MongoClient.connect(mongoURL, {useNewUrlParser: true}, function (err, client) {
-        if (err) {
-            console.log('Unable to connect to MongoDB');
-            throw err;
-        } else {
-            database.registerUser(client.db('smartmirror'), req.body, res, client);
-        }
-    });
-});
-
-// Login, check if credentials are correct and send back access_token
-app.post('/native/signin', (req, res) => {
-    MongoClient.connect(mongoURL, {useNewUrlParser: true}, function (err, client) {
-        if (err) {
-            console.log('Unable to connect to MongoDB');
-            throw err;
-        } else {
-            database.signInUser(client.db('smartmirror'), req.body, res, client);
-        }
-    });
-});
-
-// Login, check if credentials are correct and send back access_token
-app.post('/api/user/setUserWidgets', (req, res) => {
-    MongoClient.connect(mongoURL, {useNewUrlParser: true}, function (err, client) {
-        if (err) {
-            console.log('Unable to connect to MongoDB');
-            throw err;
-        } else {
-            database.setUserWidgets(client.db('smartmirror'), req.body, res, client);
-        }
-    });
-});
-
-// Get user data
-app.get('/native/getUserData', verifyToken, (req, res) => {
-    MongoClient.connect(mongoURL, {useNewUrlParser: true}, function (err, client) {
-        if (err) {
-            console.log('Unable to connect to MongoDB');
-            throw err;
-        } else {
-            jwt.verify(req.token, process.env.secretkey, (err, authData) => {
-                if (err) {
-                    res.json({});
-                } else {
-                    const userid = authData.userid;
-                    database.getUserDataForCurrentUser(client.db('smartmirror'), res, userid, client);
-                }
-            });
-        }
-    });
-});
-
-// Get user data
-app.get('/native/getAllWidgets', verifyToken, (req, res) => {
-    MongoClient.connect(mongoURL, {useNewUrlParser: true}, function (err, client) {
-        if (err) {
-            console.log('Unable to connect to MongoDB');
-            throw err;
-        } else {
-            database.getAllWidgets(client.db('smartmirror'), res, client);
-        }
-    });
-});
-
-// Get user data
-app.get('/native/getUserWidgets', verifyToken, (req, res) => {
-    MongoClient.connect(mongoURL, {useNewUrlParser: true}, function (err, client) {
-        if (err) {
-            console.log('Unable to connect to MongoDB');
-            throw err;
-        } else {
-            database.processGetUserWidgets(client.db('smartmirror'), req.query.user_id, res, client);
-        }
-    });
-});
-
-// Uploading Image for Open CV
-app.post('/native/uploadImage', verifyToken, upload.single('file'), (req, res) => {
-    if (!req.file) {
-        res.send(JSON.stringify({
-            status: false,
-            message: "Image could not be uploaded. Please try again!"
-        }));
-    } else {
-        jwt.verify(req.token, process.env.secretkey, (err, authData) => {
-            if (err) {
-                res.json({
-                    status: false,
-                    message: "User is not authorized uploading images. Please reload the application and try again!"
-                });
-            } else {
-                MongoClient.connect(mongoURL, {useNewUrlParser: true}, function (err, client) {
-                    if (err) {
-                        console.log('Unable to connect to MongoDB');
-                        res.send(JSON.stringify({
-                            status: false,
-                            message: "Database error! Please contact administrator or try again!"
-                        }));
-                    } else {
-                        const fileData = req.file;
-                        const userId = authData.userid;
-                        database.uploadImageToServer(client.db('smartmirror'), res, fileData, userId, client);
-                    }
-                });
-
-            }
-        });
-    }
-});
-
-// Upload Wunderlist Settings and clientid
-app.post('/native/uploadWunderlistSettings', verifyToken, (req, res) => {
-    jwt.verify(req.token, process.env.secretkey, (err, authData) => {
-        if (err) {
-            res.json({
-                status: false,
-                message: "User is not authorized. Please reload the application and try again!"
-            });
-        } else {
-            MongoClient.connect(mongoURL, {useNewUrlParser: true}, function (err, client) {
-                if (err) {
-                    console.log('Unable to connect to MongoDB');
-                    res.send(JSON.stringify({
-                        status: false,
-                        message: "Database error! Please contact administrator or try again!"
-                    }));
-                } else {
-                    const userId = authData.userid;
-                    database.uploadWunderlistSettings(client.db('smartmirror'), req.body, res, userId, client);
-                }
-            });
-
-        }
-    });
-});
-
-// Getting Wunderlist Settings and clientid
-app.post('/native/getWunderlistSettings', verifyToken, (req, res) => {
-    jwt.verify(req.token, process.env.secretkey, (err, authData) => {
-        if (err) {
-            res.json({
-                status: false,
-                message: "User is not authorized. Please reload the application and try again!"
-            });
-        } else {
-            MongoClient.connect(mongoURL, {useNewUrlParser: true}, function (err, client) {
-                if (err) {
-                    console.log('Unable to connect to MongoDB');
-                    res.send(JSON.stringify({
-                        status: false,
-                        message: "Database error! Please contact administrator or try again!"
-                    }));
-                } else {
-                    const userId = authData.userid;
-                    database.getWunderlistSettings(client.db('smartmirror'), res, userId, client);
-                }
-            });
-
-        }
-    });
-});
-
-
-//
-//
-//
-//
-//
-//
+app.use('/api', apiRouter);
+app.use('/native', nativeRouter);
 
 
 // Web Sockets
@@ -300,20 +62,14 @@ io.on('connection', function (socket) {
     });
 
     socket.on('app_delete_event', async function (data) {
-        MongoClient.connect(mongoURL, {useNewUrlParser: true}, async function (err, client) {
+        jwt.verify(data.token, process.env.secretkey, async (err, authData) => {
             if (err) {
-                console.log('Unable to connect to MongoDB');
                 throw err;
             } else {
-                jwt.verify(data.token, process.env.secretkey, async (err, authData) => {
-                    if (err) {
-                        throw err;
-                    } else {
-                        data.user_id = authData.username;
-                        const response = await database.removeUserWidgets(client.db('smartmirror'), data, client);
-                        io.emit('web_delete_event', data);
-                    }
-                });
+                data.user_id = authData.username;
+                const response = await userWidgetsCollectionUtils.removeUserWidgets(data);
+                console.log(response);
+                io.emit('web_delete_event', data);
             }
         });
